requirements.txt
attrs==23.2.0
black==24.4.2
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
h11==0.14.0
html5lib==1.1
idna==3.7
mechanize==0.4.10
mypy-extensions==1.0.0
outcome==1.3.0.post0
packaging==24.0
pathspec==0.12.1
platformdirs==4.2.2
PySocks==1.7.1
python-dotenv==1.0.1
requests==2.32.2
selenium==4.21.0
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
trio==0.25.1
trio-websocket==0.11.1
typing_extensions==4.12.0
urllib3==2.2.1
webdriver-manager==4.0.1
webencodings==0.5.1
wsproto==1.2.0
